% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcgwlreports-pkg.R
\docType{package}
\name{bcgwlreports-package}
\alias{bcgwlreports-package}
\alias{bcgwlreports}
\title{BC Governement Groundwater Level Reports}
\description{
Create reproducible reports of current groundwater statuses.
}
