% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additonal_functions.R
\name{get_obs_in_area}
\alias{get_obs_in_area}
\title{Get a list of all wells from a natural resource region or area.}
\usage{
get_obs_in_area(
  nr_area = c("North Natural Resource Area", "South Coast Region",
    "South Natural Resource Area", "West Coast Region"),
  rm_well = NA
)
}
\arguments{
\item{nr_area}{List one or multiple of c("North Natural Resource Area", "South Coast Region",
"South Natural Resource Area", "West Coast Region")}

\item{rm_well}{Exclude specific wells.}
}
\description{
Get a list of all wells from a natural resource region or area.
}
