% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additonal_functions.R
\name{gw_both_plots}
\alias{gw_both_plots}
\title{Create both percentiles and historical plot with selected observation wells}
\usage{
gw_both_plots(data, ows = NA)
}
\arguments{
\item{data}{Data object created from \code{gw_data_prep()} function.}

\item{ows}{List a specific well to plot from original listed wells (exports just this well)}
}
\description{
Create both percentiles and historical plot with selected observation wells
}
