% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_conditions_save.R
\name{gw_conditions_save}
\alias{gw_conditions_save}
\title{Create a percentiles map with selected observation wells}
\usage{
gw_conditions_save(
  ows = NA,
  report_dates = Sys.Date(),
  n_days = 14,
  years_min = 5,
  cache_age = 7,
  save_plots = FALSE,
  csv_dir = "",
  plots_dir = ""
)
}
\arguments{
\item{ows}{Character vector. Observation well numbers (e.g, "OW000"). Set to NA for all wells (default).}

\item{report_dates}{Character vector. Only one date for this function. Default Sys.Date().}

\item{n_days}{Numeric. If there is no data on the report date chosen, this is
the range of days over which to look for alternative dates with data.
Defaults to 2 weeks, meaning 2 weeks before and 2 weeks after a given
report date, for a total window of 4 weeks.}

\item{years_min}{Numeric. Minimum number of years required to to calculate a
percentiles}

\item{cache_age}{Logical. Maximum age in days of cached datasets (not obs well
data, but metadata related to regional maps, aquifer and wells).}
}
\description{
Create a percentiles map with selected observation wells
}
