% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additonal_functions.R
\name{gw_historic_data_plot}
\alias{gw_historic_data_plot}
\title{Create historical data plot with selected observation wells}
\usage{
gw_historic_data_plot(data, ows = NA)
}
\arguments{
\item{data}{Data object created from \code{gw_data_prep()} function.}

\item{ows}{List a specific well to plot from original listed wells (exports just this well)}
}
\description{
Create historical data plot with selected observation wells
}
