% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additonal_functions.R
\name{gw_percentile_class_table}
\alias{gw_percentile_class_table}
\title{Create a percentiles class table with selected observation wells}
\usage{
gw_percentile_class_table(data, gt = TRUE)
}
\arguments{
\item{data}{Data object created from \code{gw_data_prep()} function.}

\item{gt}{Make the table gt format (TRUE) or a regular data.frame (FALSE)}
}
\description{
Create a percentiles class table with selected observation wells
}
