% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additonal_functions.R
\name{gw_percentile_map}
\alias{gw_percentile_map}
\title{Create a percentiles map with selected observation wells}
\usage{
gw_percentile_map(data)
}
\arguments{
\item{data}{Data object created from \code{gw_data_prep()} function.}
}
\description{
Create a percentiles map with selected observation wells
}
