% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additonal_functions.R
\name{gw_wells_below_normal_table}
\alias{gw_wells_below_normal_table}
\title{Create a wells below normal table with selected observation wells}
\usage{
gw_wells_below_normal_table(
  data,
  which = c("totals", "hydraulic_connectivity", "type")[1],
  gt = TRUE
)
}
\arguments{
\item{data}{Data object created from \code{gw_data_prep()} function.}

\item{which}{Which group to filter by: "totals" for all, 'hydraulic_connectivity' or 'type' for aquifer type.}

\item{gt}{Make the table gt format (TRUE) or a regular data.frame (FALSE)}
}
\description{
Create a wells below normal table with selected observation wells
}
