% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{well_report}
\alias{well_report}
\title{Compile report}
\usage{
well_report(
  ows,
  report_dates = c(Sys.Date() - lubridate::weeks(2), Sys.Date() - lubridate::weeks(4)),
  title = NULL,
  description = NULL,
  n_days = 14,
  years_min = 5,
  out_dir = ".",
  cache_age = 7
)
}
\arguments{
\item{ows}{Character vector. Observation well numbers (e.g, "OW000")}

\item{report_dates}{Character vector. Two current dates to explore. By
default a date 2 week ago and 4 weeks before that are used.}

\item{n_days}{Numeric. If there is no data on the report date chosen, this is
the range of days over which to look for alternative dates with data.
Defaults to 2 weeks, meaning 2 weeks before and 2 weeks after a given
report date, for a total window of 4 weeks.}

\item{years_min}{Numeric. Minimum number of years required to to calculate a
percentiles}

\item{out_dir}{Character. Location of output report. Defaults to working
directory.}

\item{cache_age}{Logical. Maxmum age in days of cached datasets (not obs well
data, but metadata related to regional maps, aquifer and wells).}
}
\description{
Compile report
}
\examples{

\dontrun{

well_report(ows = c("OW008", "OW217", "OW377", "OW197"))

}

}
