% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{well_report}
\alias{well_report}
\title{Compile report}
\usage{
well_report(
  ows,
  report_dates = Sys.Date(),
  within = 7,
  years_min = 5,
  years_max = 10,
  out_dir = ".",
  cache_age = 7
)
}
\arguments{
\item{ows}{Character vector. Observation well numbers (e.g, "OW000")}

\item{report_dates}{Character vector. Two current dates to explore}

\item{within}{Numeric. Number of days within which to get alternative dates}

\item{years_min}{Numeric. Minimum number of years required to to calculate a
percentiles}

\item{years_max}{Numeric. Maximum number of years used to to calculate a
percentiles (starting with previous year).}

\item{out_dir}{Character. Location of output report. Defaults to working directory.}

\item{cache_age}{Logical. Maxmum age in days of cached datasets (not obs well
data, but metadata related to regional maps, aquifer and wells).}
}
\description{
Compile report
}
\examples{

\dontrun{

well_report(ows = c("OW008", "OW217", "OW377", "OW197"),
            report_dates = c(Sys.Date(), Sys.Date() - lubridate::weeks(2)))

}

}
